Created decorators for handling commands and onkey presses. Moved everything out of one file. Created utils file. Made a custom Terminal that extends Terminal called IntelliTerminal. Didn't mean for this to all be one commit oopsy
